+++ decidim-assemblies/lib/decidim/assemblies/participatory_space.rb
@@ -165,6 +165,20 @@ Decidim.register_participatory_space(:assemblies) do |participatory_space|
           )
         end
 
+        Decidim::AssemblyMember::POSITIONS.each do |position|
+          Decidim::AssemblyMember.create!(
+            full_name: Faker::Name.name,
+            gender: Faker::Lorem.word,
+            birthday: Faker::Date.birthday(18, 65),
+            birthplace: Faker::Demographic.demonym,
+            designation_date: Faker::Date.between(1.year.ago, 1.month.ago),
+            designation_mode: Faker::Lorem.word,
+            position: position,
+            position_other: position == "other" ? Faker::Job.position : nil,
+            assembly: current_assembly
+          )
+        end
+
         Decidim.component_manifests.each do |manifest|
           manifest.seed!(current_assembly.reload)
         end